109 meds loader (#151)
* create features for MEDS

* save to right format

* tokenization

* pretrain compatible

* added patient_info

* outcomes

* MEDS data to parquet

* better adm id assignment

* select cohort

* finetune

* remove patient_info from config

* remove iteration of hidden files

* Azure logging (#137)

### Summary of Changes:

#### **MLFlow Logging**  
- Added test logging and autologging with MLFlow.  
- Integrated logging into the trainer.  
- Improved logging setup, including logging final features and additional logging methods.  

#### **Azure Integration**  
- Added support for sub-directories/tabs for metrics in Azure.  
- Initially disabled MLFlow logging when Azure was unavailable, later removed this check as Azure is not on the cluster.  
- Added MLFlow to the CoreBEHRT Azure environment.  

#### **Refactoring & Bug Fixes**  
- Refactored components to use a new wrapper.  
- Fixed issues with the context manager in non-Azure contexts.  
- Corrected missing `}` and other minor fixes.  

#### **Metrics & Visualization**  
- Logged metrics for the best fine-tuned model.  
- Attempted to log tables, reverted to using `log_metrics` for tables.  
- Plotted end scores to sub-directories.  

#### **Miscellaneous**  
- Small refactors, utility additions, and lint fixes.  
- Added comments and improved clarity in logging functions.

* fix: log best metric only if test set present (#140)

fix: log best metric only if test set present\n\n rename to test for clarity

* Update issue templates (#135)

* 99 out of time (#147)

* feat: add cutoff option for pretrain.

* tests: add cutoff date to example config

* fix: pass absolute index date correctly

* test: pipeline test for absolute index date

* tests: update unittest for absoulute index dates

* feat: add index date shift to test set

* docs: better readme for out of time

* refactor: more explicit arg passing

* tests: new pipeline test for ft

* tests: fix typo

* fix: test_split arg parsing

* feat: introduce evaluate param

* feat: no exclusion of test_pids from folds, issue warning instead

* docs: expanded explanation of out of time evaluation

* updated tests

* added new creators test

* updated main tests

* changed configs

* black + lint

* black

* rename configs

* repartition before training

* lint

* split prepare training data from training scripts

* update azure path

* prepare pretrain to pandas

* removed dask from prepare finetune

* azure component update

* Azure sub-module updates (#152)

* Removed azure dependency

* Fixed azure config read/write with yaml

* Added pynvml to additional azure dependencies

* System metric logging

* Moved job creation to __main__

* Fix missing var

* Removed 'job' methods

* Correct name for select_cohort

* Fixed finetune naming in Azure

* Small fixes to import. Improved README

* Updated README

* Removed explicit GPU metrics logging

* Azure logging (#137)

- Added test logging and autologging with MLFlow.
- Integrated logging into the trainer.
- Improved logging setup, including logging final features and additional logging methods.

- Added support for sub-directories/tabs for metrics in Azure.
- Initially disabled MLFlow logging when Azure was unavailable, later removed this check as Azure is not on the cluster.
- Added MLFlow to the CoreBEHRT Azure environment.

- Refactored components to use a new wrapper.
- Fixed issues with the context manager in non-Azure contexts.
- Corrected missing `}` and other minor fixes.

- Logged metrics for the best fine-tuned model.
- Attempted to log tables, reverted to using `log_metrics` for tables.
- Plotted end scores to sub-directories.

- Small refactors, utility additions, and lint fixes.
- Added comments and improved clarity in logging functions.

* fix: log best metric only if test set present (#140)

fix: log best metric only if test set present\n\n rename to test for clarity

* 99 out of time (#147)

* feat: add cutoff option for pretrain.

* tests: add cutoff date to example config

* fix: pass absolute index date correctly

* test: pipeline test for absolute index date

* tests: update unittest for absoulute index dates

* feat: add index date shift to test set

* docs: better readme for out of time

* refactor: more explicit arg passing

* tests: new pipeline test for ft

* tests: fix typo

* fix: test_split arg parsing

* feat: introduce evaluate param

* feat: no exclusion of test_pids from folds, issue warning instead

* docs: expanded explanation of out of time evaluation

* prepare pretrain to pandas

* azure component update

* compatible with new azure updates

* updates tests

* lint

* format

* included prepare_data in pipeline

* added function in components

* fix pipeline

* black

* tests: fix pipeline yaml

* finetune component

* updated finetune configs

* added cohort to configs

* split logic moved to prepare for pretrain

* added test for prepare_pretrain

* split up background creation and patient_info creation

* removed _sort_partition

* moved get_adm function

* make held_out set optional

* remove Excluder

* subject_id to ints

* value creation using helper column

* cleaner adm_id creation

* added constants

* decoupling of admissions and segments

* updated readme

* format

* lint

* removed cutoffs in config

* removed parquetloader

* lint

---------

+++ corebehrt/main/create_data.py
@@ -12,14 +12,10 @@ import logging
 import os
 from os.path import join
 
-import dask.dataframe as dd
 import torch
-from dask.diagnostics import ProgressBar
 
-from corebehrt.functional.features.split import split_pids_into_pt_ft_test
 from corebehrt.functional.io_operations.load import load_vocabulary
 from corebehrt.functional.setup.args import get_args
-from corebehrt.modules.features.excluder import Excluder
 from corebehrt.modules.features.tokenizer import EHRTokenizer
 from corebehrt.modules.setup.config import load_config
 from corebehrt.modules.setup.directory import DirectoryPreparer
@@ -53,26 +49,16 @@ def main_data(config_path):
     # TODO: temporary fix/check until we split the script into two.
     # As cfg.paths.features is always set, its value cannot be used to decide
     # if features are present.
-    if os.path.exists(join(cfg.paths.features, "part.0.parquet")):
+    if os.path.exists(join(cfg.paths.features, "held_out", "0.parquet")):
         logger.info("Reusing existing features")
     else:
         logger.info("Create and process features")
-        create_and_save_features(
-            Excluder(**cfg.excluder),  # Excluder is the new Handler and old Excluder
-            cfg,
-        )
+        create_and_save_features(cfg)
         logger.info("Finished feature creation and processing")
 
-    logger.info("Get all pids")
-    df = dd.read_parquet(cfg.paths.features)
-    pids = df.PID.unique().compute().tolist()
-
-    logger.info("Split pids")
-    pretrain_pids, finetune_pids, test_pids = split_pids_into_pt_ft_test(
-        pids, **cfg.split_ratios
-    )
-
     logger.info("Tokenizing")
+    features_path = cfg.paths.features
+    tokenized_path = cfg.paths.tokenized
 
     vocabulary = None
     if "vocabulary" in cfg.paths:
@@ -80,26 +66,23 @@ def main_data(config_path):
         vocabulary = load_vocabulary(cfg.paths.vocabulary)
     tokenizer = EHRTokenizer(vocabulary=vocabulary, **cfg.tokenizer)
 
-    features_path = cfg.paths.features
-    tokenized_path = cfg.paths.tokenized
+    logger.info("Tokenizing train")
+    load_tokenize_and_save(
+        features_path,
+        tokenizer,
+        tokenized_path,
+        "train",
+    )
+    tokenizer.freeze_vocabulary()
 
-    with ProgressBar(dt=1):
-        logger.info("Tokenizing pretrain")
-        load_tokenize_and_save(
-            features_path, tokenizer, tokenized_path, "pretrain", pretrain_pids
-        )
-        tokenizer.freeze_vocabulary()
-        logger.info("Tokenizing finetune")
-        load_tokenize_and_save(
-            features_path, tokenizer, tokenized_path, "finetune", finetune_pids
-        )
-        logger.info("Tokenizing test")
-        load_tokenize_and_save(
-            features_path, tokenizer, tokenized_path, "test", test_pids
-        )
+    logger.info("Tokenizing tuning")
+    load_tokenize_and_save(features_path, tokenizer, tokenized_path, "tuning")
 
-    torch.save(tokenizer.vocabulary, join(tokenized_path, "vocabulary.pt"))
-    logger.info("Finished tokenizing")
+    if os.path.exists(os.path.join(features_path, "held_out")):
+        logger.info("Tokenizing held_out")
+        load_tokenize_and_save(features_path, tokenizer, tokenized_path, "held_out")
+        torch.save(tokenizer.vocabulary, join(tokenized_path, "vocabulary.pt"))
+        logger.info("Finished tokenizing")
 
 
 if __name__ == "__main__":
